Server/requirements.txt
Flask
google-cloud
apache-libcloud
pycrypto
gunicorn
oauth2client
flask-httpauth
itsdangerous
